Frontend: Refactor login mechanism to use localStorage for the time being
THIS IS NOT SECURE since the player id is basically the access token. Later we will store a JWT instead
